+++ src/app/app.component.ts
@@ -1,12 +1,10 @@
 import { Component } from '@angular/core';
-
 import { Platform } from '@ionic/angular';
 import { AuthServiceProvider } from './services/security/auth-service';
 import { TranslateService } from '@ngx-translate/core';
 import { Router } from '@angular/router';
 
 import { Plugins, Capacitor } from '@capacitor/core';
-
 const { SplashScreen, StatusBar } = Plugins;
 
 @Component({
