+++ src/app/pages/home/home.module.ts
@@ -1,5 +1,5 @@
 import { NgModule, Component } from '@angular/core';
-import { HomePage } from './home';
+import { HomePage } from './home.page';
 import { IonicModule } from '@ionic/angular';
 import { CommonModule } from '@angular/common';
 import { Routes, RouterModule } from '@angular/router';
