+++ src/app/pages/home/home.page.ts
@@ -11,8 +11,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'page-Home',
-  templateUrl: 'home.html',
-  styleUrls: ['home.scss'],
+  templateUrl: 'home.page.html',
+  styleUrls: ['home.page.scss'],
 })
 export class HomePage {
