+++ src/app/pages/sampledata-detail/sampledata-detail.page.html
