+++ src/app/pages/sampledata-detail/sampledata-detail.page.ts
@@ -14,8 +14,8 @@ import { PaginatedListTo } from '../../services/interfaces/paginated-list-to';
  */
 @Component({
   selector: 'sampledata-detail',
-  templateUrl: 'sampledata-detail.html',
-  styleUrls: ['sampledata-detail.scss']
+  templateUrl: 'sampledata-detail.page.html',
+  styleUrls: ['sampledata-detail.page.scss']
 })
 export class SampledataDetail {
   pageable: Pageable = {
